EntityCheck builder split & Messages added to builder
