+++ IDesign/IDesign.Regonizers/Models/Checks/Entities/GroupChecks/AllMemberGroupCheck.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using IDesign.Recognizers.Abstractions;
+using IDesign.Recognizers.Models.Output;
+using SyntaxTree.Abstractions.Entities;
+using SyntaxTree.Abstractions.Members;
+
+namespace IDesign.Recognizers.Models.Checks.Entities.GroupChecks
+{
+    public class AllMemberGroupCheck : MemberGroupCheck
+    {
+        public AllMemberGroupCheck(ICheckBuilder<IEntity> root) : base(root)
+        {
+        }
+
+        public override ICheckResult Check(Dictionary<IMember, List<ICheckResult>> results, IMemberCheckBuilder check)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
